pyproject.toml
[tool.poetry]
name = "twm"
version = "0.1.0"
description = ""
authors = ["wassname <git@wassname.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
torch = {version = "^2.1.0+cu121", source = "pytorch"}
# pandas = "^2.2.2"
# matplotlib = "^3.8.0"
loguru = "^0.7.2"
# einops = "^0.8.0"
# torchinfo = "^1.8.0"
# accelerate = "^0.24.1"
# peft = "^0.5.0"
# bitsandbytes = "^0.41.0"
# transformers = "4.34.0"
tqdm = "^4.66.1"
wandb = "^0.12.6"
pygame = "^2.1.0"
torchvision = "^0.16.0"
# numpy = ">=1.18.0"
# scipy = "^1.11.3"
craftax = {path = "/media/wassname/SGIronWolf/projects5/2024/Craftax"}
Pillow = "^9.0.0"
# ale_py = "^0.7.0"
protobuf = "^3.10.0"
gymnasium = {version = "^0.29.1", extras = ["atari", "accept-rom-license", "other"]}
jaxtyping = "^0.2.2"
# typeguard = "^4.2.20"
beartype = "^0.18.5"
jax = {version = "^0.4.28", source = "jax", extras = ["cuda12_pip"]}

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[[tool.poetry.source]]
name = "jax"
url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
priority = "explicit"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"
ruff = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
